And share_text, toast, recreate
